Code review 21/10/2022
