--- common/src/main/java/com/dmdev/entity/presets/Importance.java
@@ -1,9 +1,0 @@
-package com.dmdev.entity.presets;
-
-public enum Importance {
-    LEVEL_ZERO,
-    LEVEL_ONE,
-    LEVEL_TWO,
-    LEVEL_THREE,
-    LEVEL_FOUR
-}
