--- common/src/main/java/com/dmdev/entity/presets/Role.java
@@ -1,5 +1,0 @@
-package com.dmdev.entity.presets;
-
-public enum Role {
-    USER, ADMIN
-}
