+++ service/src/main/java/entity/Gift.java
@@ -1,0 +1,53 @@
+package entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToMany;
+import javax.persistence.ManyToOne;
+import java.time.LocalDate;
+import java.util.HashSet;
+import java.util.Set;
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@ToString(exclude = {"person", "tags"})
+@EqualsAndHashCode(exclude = {"person", "tags"})
+@Entity
+public class Gift {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String name;
+    private String description;
+    private LocalDate date;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Person person;
+
+    @Builder.Default
+    @JoinTable(
+            name = "gift_tag",
+            joinColumns = @JoinColumn(name = "gift_id"),
+            inverseJoinColumns = @JoinColumn(name = "tag_id"))
+    @ManyToMany
+    private Set<Tag> tags = new HashSet<>();
+
+    public void addTag(Tag tag) {
+        tags.add(tag);
+        tag.getGifts().add(this);
+    }
+}
