+++ service/src/main/java/entity/Scope.java
@@ -1,0 +1,34 @@
+package entity;
+
+import entity.profiles.ProfessionalProfile;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@ToString(exclude = "professionalProfile")
+@EqualsAndHashCode(exclude = "professionalProfile")
+@Entity
+public class Scope {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String name;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private ProfessionalProfile professionalProfile;
+}
