+++ service/src/main/java/entity/presets/Importance.java
@@ -1,0 +1,9 @@
+package entity.presets;
+
+public enum Importance {
+    LEVEL_ZERO,
+    LEVEL_ONE,
+    LEVEL_TWO,
+    LEVEL_THREE,
+    LEVEL_FOUR
+}
