+++ service/src/main/java/entity/presets/Role.java
@@ -1,0 +1,5 @@
+package entity.presets;
+
+public enum Role {
+    USER, ADMIN
+}
