+++ service/src/main/java/entity/profiles/PrivateProfile.java
@@ -1,0 +1,25 @@
+package entity.profiles;
+
+import entity.Person;
+import lombok.*;
+
+import javax.persistence.*;
+import java.time.LocalDate;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(exclude = "person")
+@Builder
+@Table(name = "private_profile")
+public class PrivateProfile {
+
+    @Id
+    private Integer id;
+    private LocalDate wedding;
+
+    @PrimaryKeyJoinColumn
+    @OneToOne
+    private Person person;
+}
