+++ service/src/main/resources/project.description
@@ -1,0 +1,30 @@
+1.Проект под рабочим названием "Storyteller"
+1.1 Заказчик - мастерская эксклюзивного переплета, работающая на 80% в формате b2g.
+1.2 Клиенты (users) - сотрудники госструктур, отвечающие за подарки чиновникам и протокол.
+
+1.3 Цель проекта: систематизировать информацию о событиях (stories) людей (persons) , которых нужно поздравлять.
+Чтобы предупредить ситуации по судорожному изготовлению подарка в кратчайшие сроки
+
+2.1 Профиль человека делится на три категории: общий (personal_profile), рабочий (professional_profile) и личный (private_profile)
+2.2 Каждый человек в системе имеет коэффициент важности.
+
+3.1 Все события имеют тег (госпраздник, религиозный праздник, свадьба, ДР и пр.) и принадлежность к человеку
+
+4. Второй модуль системы - реестр подарков.
+4.1 Каждый подарок имеет свою ценность и тег (условно, Коран для мусульман, Библия для христиан и т.п.)
+4.2 За каждым подарком фиксируются данные, кому и когда он уже дарился
+
+5. Порядок работы и монетизация
+5.1 Клиент получает/покупает доступ/подписку к тегу (например, МинОбразования),
+получая таким образом информацию только о данном учреждении.
+5.2 Мастерская заранее знает о предстоящих событиях и может подготовить подарки и напомнить клиенту о необходимости их приобретения.
+5.3 Партнерские мастерские могут предложить свою продукцию для включения ее в реестр подарков
+5.4 Заказчик может иметь комиссию с заказов из п.5.3
+
+6. Перспективы
+6.1 Клиенту приходит уведомление о предстоящем событии (ДР начальника, день собственной свадьбы и т.п.),
+после чего ему предлагаются на выбор несколько подарков, исходя из тематики события
+и его личного профиля (история заказов, личная информация, предпочтения). Заказ происходит в один клик, избавляя от
+необходимости поиска и подбора.
+
+p.S. Система позволит исключить вопрос "Дорогой, а ты помнишь, какой сегодня день?". Все будет автоматизировано.
