+++ service/src/test/resources/hibernate.cfg.xml
@@ -4,7 +4,7 @@
         "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
 <hibernate-configuration>
     <session-factory>
-        <property name="connection.url">jdbc:postgresql://localhost:5432/storyteller</property>
+        <property name="connection.url">jdbc:postgresql://localhost:5432/test-storyteller</property>
         <property name="connection.username">postgres</property>
         <property name="connection.password">Masterhost25</property>
         <property name="connection.driver_class">org.postgresql.Driver</property>
